+++ lib/spack/spack/cmd/dependencies.py
@@ -11,8 +11,7 @@ import spack.cmd
 import spack.environment as ev
 import spack.store
 from spack.cmd.common import arguments
-from spack.solver.context import create_inspector, default_context
-from spack.solver.counter import PossibleDependenciesAnalyzer
+from spack.solver.context import default_inspector
 
 description = "show dependencies of a package"
 section = "basic"
@@ -69,9 +68,7 @@ def dependencies(parser, args):
 
     else:
         spec = specs[0]
-        dependencies, virtuals, _ = PossibleDependenciesAnalyzer(
-            create_inspector(default_context())
-        ).possible_dependencies(
+        dependencies, virtuals, _ = default_inspector().possible_dependencies(
             spec,
             transitive=args.transitive,
             expand_virtuals=args.expand_virtuals,
